Check status once again before committing
The reason for that is that some users may want to remove all
translation before downloading them again, for instance, to make sure
there are no stale translations.

When that is the case, the first check will indicate there are changes,
since files were removed.

However, when the download files are the same as the files that were
previously removed, `git add .` will cause the next `git status -s` to
be clean. If the second check is not in place, the script will attempt
to commit the changes, but as there are none, the action will fail,
instead of just being skipped.

+++ entrypoint.sh
@@ -153,6 +153,12 @@ push_to_branch() {
     echo "PUSH TO BRANCH ${LOCALIZATION_BRANCH}"
 
     git add .
+
+    if [ ! -n "$(git status -s)" ]; then
+      echo "NOTHING TO COMMIT"
+      return
+    fi
+
     git commit --no-verify -m "${INPUT_COMMIT_MESSAGE}"
     git push --no-verify --force "${REPO_URL}"
